renamed all files to be lowercase and updated paths
